requirements.txt
asgiref==3.2.7
backcall==0.1.0
boto3==1.13.18
botocore==1.16.18
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.42.0
django-cors-headers==3.3.0
django-resized==0.3.11
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.16
Faker==4.1.0
gunicorn==20.0.4
idna==2.9
ipdb==0.13.2
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
jmespath==0.10.0
mailchimp3==3.0.14
oauthlib==3.1.0
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
psycopg2==2.8.5
ptyprocess==0.6.0
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
simplegeneric==0.8.1
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
whitenoise==5.1.0
